+++ service/src/main/kotlin/io/provenance/explorer/service/AssetService.kt
@@ -2,13 +2,22 @@ package io.provenance.explorer.service
 
 import com.google.protobuf.Timestamp
 import com.google.protobuf.util.JsonFormat
+import io.ktor.client.call.receive
+import io.ktor.client.features.ResponseException
+import io.ktor.client.request.get
+import io.ktor.client.request.parameter
+import io.ktor.client.statement.HttpResponse
+import io.provenance.explorer.KTOR_CLIENT
+import io.provenance.explorer.config.ExplorerProperties
 import io.provenance.explorer.config.ResourceNotFoundException
 import io.provenance.explorer.domain.core.logger
 import io.provenance.explorer.domain.entities.BaseDenomType
 import io.provenance.explorer.domain.entities.MarkerCacheRecord
+import io.provenance.explorer.domain.entities.MarkerCacheTable
 import io.provenance.explorer.domain.entities.TokenDistributionAmountsRecord
 import io.provenance.explorer.domain.entities.TokenDistributionPaginatedResultsRecord
 import io.provenance.explorer.domain.entities.TxMarkerJoinRecord
+import io.provenance.explorer.domain.extensions.NHASH
 import io.provenance.explorer.domain.extensions.pageCountOfResults
 import io.provenance.explorer.domain.extensions.toDateTime
 import io.provenance.explorer.domain.extensions.toObjectNode
@@ -17,14 +26,15 @@ import io.provenance.explorer.domain.models.explorer.AssetDetail
 import io.provenance.explorer.domain.models.explorer.AssetHolder
 import io.provenance.explorer.domain.models.explorer.AssetListed
 import io.provenance.explorer.domain.models.explorer.AssetManagement
-import io.provenance.explorer.domain.models.explorer.CoinStr
 import io.provenance.explorer.domain.models.explorer.CountStrTotal
 import io.provenance.explorer.domain.models.explorer.PagedResults
 import io.provenance.explorer.domain.models.explorer.TokenCounts
 import io.provenance.explorer.domain.models.explorer.TokenDistribution
 import io.provenance.explorer.domain.models.explorer.TokenDistributionAmount
+import io.provenance.explorer.domain.models.explorer.toCoinStrWithPrice
 import io.provenance.explorer.grpc.extensions.getManagingAccounts
 import io.provenance.explorer.grpc.extensions.isMintable
+import io.provenance.explorer.grpc.v1.AccountGrpcClient
 import io.provenance.explorer.grpc.v1.AttributeGrpcClient
 import io.provenance.explorer.grpc.v1.MarkerGrpcClient
 import io.provenance.explorer.grpc.v1.MetadataGrpcClient
@@ -34,7 +44,10 @@ import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.runBlocking
+import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.transactions.transaction
+import org.json.JSONArray
+import org.json.JSONObject
 import org.springframework.stereotype.Service
 import java.math.BigDecimal
 import java.math.RoundingMode
@@ -44,8 +57,9 @@ class AssetService(
     private val markerClient: MarkerGrpcClient,
     private val attrClient: AttributeGrpcClient,
     private val metadataClient: MetadataGrpcClient,
-    private val accountService: AccountService,
-    private val protoPrinter: JsonFormat.Printer
+    private val accountClient: AccountGrpcClient,
+    private val protoPrinter: JsonFormat.Printer,
+    private val props: ExplorerProperties
 ) {
     protected val logger = logger(AssetService::class)
 
@@ -54,20 +68,19 @@ class AssetService(
         page: Int,
         count: Int
     ): PagedResults<AssetListed> {
-        val list =
-            MarkerCacheRecord
-                .findByStatusPaginated(statuses, page.toOffset(count), count)
-                .map {
-                    AssetListed(
-                        it.denom,
-                        it.markerAddress,
-                        CoinStr(it.supply.toBigInteger().toString(), it.denom),
-                        it.status.prettyStatus(),
-                        it.data?.isMintable() ?: false,
-                        it.lastTx?.toString(),
-                        it.markerType.prettyMarkerType()
-                    )
-                }
+        val records = MarkerCacheRecord.findByStatusPaginated(statuses, page.toOffset(count), count)
+        val pricing = getPricingInfoIn(records.map { it.denom })
+        val list = records.map {
+            AssetListed(
+                it.denom,
+                it.markerAddress,
+                it.toCoinStrWithPrice(pricing[it.denom]),
+                it.status.prettyStatus(),
+                it.data?.isMintable() ?: false,
+                it.lastTx?.toString(),
+                it.markerType.prettyMarkerType()
+            )
+        }
         val total = MarkerCacheRecord.findCountByStatus(statuses)
         return PagedResults(total.pageCountOfResults(count), list, total)
     }
@@ -88,10 +101,10 @@ class AssetService(
                 it.denom,
                 it.status.toString(),
                 it,
-                accountService.getCurrentSupply(denom).toBigDecimal(),
+                getCurrentSupply(denom).toBigDecimal(),
                 TxMarkerJoinRecord.findLatestTxByDenom(denom)
             )
-        } ?: accountService.getCurrentSupply(denom).let {
+        } ?: getCurrentSupply(denom).let {
             MarkerCacheRecord.insertIgnore(
                 null,
                 denom.getBaseDenomType().name,
@@ -109,7 +122,8 @@ class AssetService(
             getAssetFromDB(denom)?.let { (id, record) ->
                 val txCount = TxMarkerJoinRecord.findCountByDenom(id.value)
                 val attributes = async { attrClient.getAllAttributesForAddress(record.markerAddress) }
-                val balances = async { accountService.getBalances(record.markerAddress!!, 1, 1) }
+                val balances = async { accountClient.getAccountBalances(record.markerAddress!!, 1, 1) }
+                val price = getPricingInfoIn(listOf(denom))[denom]
                 AssetDetail(
                     record.denom,
                     record.markerAddress,
@@ -117,12 +131,12 @@ class AssetService(
                         record.data!!.getManagingAccounts(),
                         record.data!!.allowGovernanceControl
                     ) else null,
-                    CoinStr(record.supply.toBigInteger().toString(), record.denom),
+                    record.toCoinStrWithPrice(price),
                     record.data?.isMintable() ?: false,
                     if (record.markerAddress != null) markerClient.getMarkerHoldersCount(denom).pagination.total.toInt() else 0,
                     txCount,
                     attributes.await().map { attr -> attr.toResponse() },
-                    accountService.getDenomMetadataSingle(denom).toObjectNode(protoPrinter),
+                    getDenomMetadataSingle(denom).toObjectNode(protoPrinter),
                     TokenCounts(
                         if (record.markerAddress != null) balances.await().pagination.total else 0,
                         if (record.markerAddress != null) metadataClient.getScopesByOwner(record.markerAddress!!).pagination.total.toInt() else 0
@@ -135,7 +149,7 @@ class AssetService(
 
     fun getAssetHolders(denom: String, page: Int, count: Int) =
         runBlocking {
-            val supply = accountService.getCurrentSupply(denom)
+            val supply = getCurrentSupply(denom)
             val res = markerClient.getMarkerHolders(denom, page.toOffset(count), count)
             val list = res.balancesList.asFlow().map { bal ->
                 val balance = bal.coinsList.first { coin -> coin.denom == denom }.amount
@@ -216,7 +230,7 @@ class AssetService(
         TokenDistributionAmountsRecord.batchUpsert(tokenDistributions)
     }
 
-    fun getMetadata(denom: String?) = accountService.getDenomMetadata(denom).map { it.toObjectNode(protoPrinter) }
+    fun getMetadata(denom: String?) = getDenomMetadata(denom).map { it.toObjectNode(protoPrinter) }
 
     // Updates the Marker cache
     fun updateAssets(denoms: Set<String>, txTime: Timestamp) =
@@ -226,13 +240,93 @@ class AssetService(
                     val data = markerClient.getMarkerDetail(marker)
                     MarkerCacheRecord.findByDenom(marker)?.apply {
                         if (data != null) this.status = data.status.toString()
-                        this.supply = accountService.getCurrentSupply(marker).toBigDecimal()
+                        this.supply = getCurrentSupply(marker).toBigDecimal()
                         this.lastTx = txTime.toDateTime()
                         this.data = data
                     }
                 }
             }
         }
+
+    fun getTotalAum() = runBlocking {
+        val baseMap = MarkerCacheRecord.find {
+            (MarkerCacheTable.status eq MarkerStatus.MARKER_STATUS_ACTIVE.name) and
+                (MarkerCacheTable.supply greater BigDecimal.ZERO)
+        }.associate { it.denom to it.supply }
+        val pricing = baseMap.keys.toList().chunked(50) { getPricingInfo(it) }.flatMap { it.toList() }.toMap()
+            .toMutableMap().also { it.putAll(getPricingForNhash()) }
+        baseMap.map { (k, v) -> (pricing[k] ?: BigDecimal.ZERO).multiply(v.toHashSupply(k)) }.sumOf { it }
+    }
+
+    fun getAumForList(denoms: Map<String, String>): BigDecimal {
+        val pricing =
+            denoms.keys.toList().chunked(50) { getPricingInfo(it) }.flatMap { it.toList() }.toMap().toMutableMap()
+                .also { it.putAll(getPricingForNhash()) }
+        return denoms
+            .map { (k, v) -> (pricing[k] ?: BigDecimal.ZERO).multiply(v.toBigDecimal().toHashSupply(k)) }
+            .sumOf { it }
+    }
+
+    fun getPricingInfoIn(denoms: List<String>) =
+        getPricingInfo(denoms).also { it.putAll(getPricingForNhash()) }
+
+    fun getPricingInfo(denoms: List<String>): MutableMap<String, BigDecimal> = runBlocking {
+        fun JSONObject.getDecimalOrNull(key: String) = try {
+            this.getBigDecimal(key)
+        } catch (e: Exception) {
+            BigDecimal.ZERO
+        }
+
+        val res = try {
+            KTOR_CLIENT.get<HttpResponse>("${props.pricingUrl}/api/v1/pricing/marker/denom/list") {
+                parameter("denom[]", denoms.joinToString(","))
+            }
+        } catch (e: ResponseException) {
+            return@runBlocking mutableMapOf<String, BigDecimal>().also { logger.error("Error: ${e.response}") }
+        }
+
+        if (res.status.value in 200..299) {
+            try {
+                JSONArray(res.receive<String>()).mapNotNull { ele ->
+                    if (ele is JSONObject) {
+                        ele.getString("markerDenom") to ele.getDecimalOrNull("usdPrice")
+                    } else null
+                }.toMap().toMutableMap()
+            } catch (e: Exception) {
+                mutableMapOf<String, BigDecimal>().also { logger.error("Error: $e") }
+            }
+        } else mutableMapOf<String, BigDecimal>()
+            .also { logger.error("Error reaching Pricing Engine: ${res.status.value}") }
+    }
+
+    fun getPricingForNhash(): Map<String, BigDecimal> = runBlocking {
+        val url =
+            "https://www.dlob.io/aggregator/external/api/v1/order-books/pb18vd8fpwxzck93qlwghaj6arh4p7c5n894vnu5g/daily-price"
+        val res = try {
+            KTOR_CLIENT.get<HttpResponse>(url)
+        } catch (e: ResponseException) {
+            return@runBlocking mapOf<String, BigDecimal>().also { logger.error("Error: ${e.response}") }
+        }
+
+        if (res.status.value == 200) {
+            try {
+                JSONObject(res.receive<String>()).let {
+                    mapOf("nhash" to it.getString("latestDisplayPricePerDisplayUnit").toBigDecimal())
+                }
+            } catch (e: Exception) {
+                mapOf<String, BigDecimal>().also { logger.error("Error: $e") }
+            }
+        } else mapOf<String, BigDecimal>().also { logger.error("Error reaching Pricing Engine: ${res.status.value}") }
+    }
+
+    fun getCurrentSupply(denom: String) = runBlocking { accountClient.getCurrentSupply(denom).amount }
+
+    fun getDenomMetadataSingle(denom: String) = runBlocking { accountClient.getDenomMetadata(denom).metadata }
+
+    fun getDenomMetadata(denom: String?) = runBlocking {
+        if (denom != null) listOf(accountClient.getDenomMetadata(denom).metadata)
+        else accountClient.getAllDenomMetadata().metadatasList
+    }
 }
 
 fun BigDecimal.asPercentOf(divisor: BigDecimal): BigDecimal = this.divide(divisor, 20, RoundingMode.CEILING)
@@ -248,3 +342,5 @@ fun String.getBaseDenomType() =
         this.startsWith("ibc/") -> BaseDenomType.IBC_DENOM
         else -> BaseDenomType.DENOM
     }
+
+fun BigDecimal.toHashSupply(denom: String) = if (denom == NHASH) this.divide(1000000000.toBigDecimal()) else this